Minor updates to the GcodeDriver console. Didn't notice that the file was in the wrong place, so I moved it, updated imports and reformatted.
